Add support for get_server_auth_methods()
This commit adds support for a new get_server_auth_methods() function
which contacts an SSH server and returns a list of available auth
methods for a given user. Authentication is not attempts - the
connection is closed as soon as the list of auth methods is retrieved.
However, for this to work, the SSH key exchange must complete
successfully.