Add notes to simple client & server examples about encoding
This commit adds a note in the simple client and server examples about
how I/O is done on strings by default, due to the default encoding of
'utf-8', and how callers can pass in an encoding of None if they want
to operate on bytes instead.

The opposite of this was previously documented down in the section on
direct and forwarded TCP connections, explaining that they default to
operating on bytes by default, but you can explicitly set the encoding
if a caller wants to send and receive strings on TCP channels.