Increase cryptography minimum version to 3.1
This commit increases the minimum required version of the Python
cryptography package to 3.1 in AsyncSSH, to account for changes made
in commit 88dcdac which stopped passing in the "backend" argument
on some cryptography calls. This argument is now deprecated, but
version 3.1 was the first release which allowed this parameter to
not be passed in.

Thanks go to Github users sebby97 and JavaScriptDude for reporting
this issue!