Fix an issue with server channels not picking up connection settings
This commit fixes an issue where channel-level settings such as
encoding, errors, window, and max_pktsize set when opening the listener
were not being properly applied when new sessions were opened by
clients.

Also, this commit updates the documentation in multiple places to
specify that encoding can be either a string or None.