Add a workaround for a bug in dropbear
This commit adds a workaround in AsyncSSH for a bug in dropbear related
to maximum decompressed payload size when compression is enabled. The
problem is an off-by-one error introduced back in 2013 in dropbear commit
https://github.com/mkj/dropbear/commit/49263b5 which can incorrectly report
that a decompressed packet is too large if the peer sends a packet which
is exactly 32768 bytes in size after decompression. The workaround reduces
the maximum send packet size to 32767 bytes when dropbear is detected and
compression is enabled to avoid triggering this error.

Thanks go to Matti Niemenmaa for reporting this issue and providing an
excellent test case to help track down the source of the problem!