Add support for config directive "Match Exec"
This commit adds support for the "Match Exec" directive in OpenSSH
config files. However, because the existing config file API uses
syncrhonous calls, this should be used with caution. With the
current implementation, the event loop will be completely blocked
during config parsing, preventing other tasks from making progress
if a "Match Exec" command doesn't return an immediate result, and
a warning to that effect has been added to the documentation.