Change coverage report to run on Python 3.7 to work around bug
When collecting coverage data on Python 3.7 and earlier but generating
the coverage report in 3.8 or later, some decorators show incorrect
coverage information (see https://github.com/nedbat/coveragepy/issues/866
for more details).

This commit changes the github action to use Python 3.7 when generating
the coverage report, hopefully working around this issue for some of the
Windows-only modules which are only run on Python 3.6 and 3.7.