Fix GSS MIC auth to work with GSS key exchange on AsyncSSH server
This commit fixes GSS MIC auth to to work in conjunction with GSS key
exchange when AsyncSSH is running as a server. This was previously
fixed in commit d06383d on the client side, but a similar fix for the
server side was not included at that time.

This only appears when GSS key exchange is allowed as a kex algorithm
but then disabled as a user authentication mechanism, which would not
be common. However, with this fix this combination should now work
corrrectly.