More glob cleanup
This commit updates glob matching of the recursive '**' pattern to
include all of the directories matching the prefix of the '**' match
in addition to any files and subdirectories under those prefixes.
This makes it more consistent with the results returned by glob.glob().

Some other minor cleanup was also done to simplify handling of cache
misses in the scandir and stat caches.