Properly mark channels as closed when receiving connection close
This commit updates the connection close handler to properly mark all
channels as fully closed when the peer unexpected closes the connection,
allowing exceptions to fire if an application continues to try and use
the channel. Thanks go to Taha Jahangir for reporting this issue and
suggesting a possible fix.