Fix an issue with readuntil() with certain characters in pattern
This commit fixes an issue with readuntil() where it didn't properly
match when certain punctuation characters were present in the string
pattern to look for. The issue appeared when characters in the string
had special meaning in regex matching, as they weren't being properly
escaped. Thanks go to Github user MazokuMaxy for reporting this issue.