Fix Pageant auth support
This commit fixes an issue where agent auth was incorrectly disabled
when no agent_path or SSH_AUTH_SOCK was provided, disabling the
fallback to Pageant auth or the default ssh-agent path on Windows.
This was an unintentional result of changes made when adding type
annotation support in release 2.9.0.

Changes here include tightening up the allowed types when calling
into the agent modules to reduce the chance of this type of problem
happening again.