Add argument to set signature algorithm when creating OpenSSH certs
This commit adds a new "sig_alg" argument to generate_user_certificate()
and generate_host_certificate() on the SSHKey class, allowing the caller
to choose which signature algorithm to use when creating the cert.

In practice, this is only needed for RSA keys, as all other keys support
only a single signature algorithm for each key type. However, for RSA
keys, it's now possible to choose between the ssh-rsa, rsa-sha2-256, and
rsa-sha2-512 signature algorithms when creating OpenSSH certificates.