Allow redirect to forward break, signal, and terminal size messages
This commit enhances the SSHProcess redirect functionality to be able
to forward break, signal, and terminal size change messages when an
SSHServerProcess receiving those messages redirects stdin to an
SSHClientProcess. With this change, those messages are now forwarded
on to the upstream server along with any data received on stdin.