Switch from listdir to scandir for SFTP/SCP
This commit switches SFTP and SCP to use scandir() instead of listdir(),
reducing the number of separate calls to stat() and allowing processing
to be done incrementally, rather than fetching the entire contents of a
directory before iterating over it.

This change maintains backward compatibility with code that provides its
own SFTPServer class with an alternate implementation of listdir().
However. the recommended approach it now to implement scandir() instead
of listdir() in custom SFTPServer classes.