Add new read_parallel() method to SFTPClientFile
This commit adds a new read_parallel() function to SFTPClientFile,
providing a way to schedule a set of parallel reads of a remote file and
get back the data a block at a time as the reads complete. Previously,
read() with a large size would do the read in parallel, but it would
wait until all the data came back and was reassembled before anything
was returned.

Since reads may complete out of order, the data is returned as an async
iterator of tuples containing the offset within the file and the data
read at that offset.

Just like the existing read() call, the block size of the individual
reads and the maximum number of parallel read requests can be controlled
using the block_size and max_requests parameters which can be set when
the file is opened.