Add "sock" argument to top-level AsyncSSH functions
This commit adds a "sock" argument to the AsyncSSH connect(),
connect_reverse(), listen(), listen_reverse(), get_server_host_key(),
and get_server_auth_methods() functions, allowing an existing
already-connected socket to be used in place of making a new outbound
connection, or allowing an existing bound socket to be used in place
of binding a new listening socket.

This argument is passed on to the asyncio.create_connection() and
asyncio.create_server() functions, and has the same restriction that
host and port information and other arguments related to connecting
or binding the socket should not be passed when sock is passed.