Ignore encrypted keys in config by default when no passphrase is given
This commit changes the default of the ignore_encrypted option to be
True when keys are being loaded from config files, automatically
ignoring encrypted keys when no passphrase is specified. For keys
set via client_keys, the default is to still give an error in this
case unless ignore_encrypted is explicitly set.

When loading keys from their default locations, encrypted keys were
already ignored when no passphrase is given, and this behavior isn't
changing.