Fix an issue with SFTP servers which don't support random access I/O
This commit fixes an issue introduced in AsyncSSH 2.12 where SFTP file
copies stopped working to servers which don't support random access I/O.

Because AsyncSSH performs parallel reads by default, there's a
possibility these reads can complete out of order in some cases,
breaking if the server fails when write requests are not in order.
This could happen on older AsyncSSH releases, but it wasn't common
and neither are servers that don't accept random I/O.

With AsyncSSH 2.12, a change inadvertantly caused out of order write
requests to become much more common, leading to this issue being
discovered. This commit doesn't guarantee in-order writes, but it
returns AsyncSSH to its 2.11 behavior with regard to this, which
should avoid this issue in most cases.

In cases where this is still a problem, setitng max_requests=1 in
the copy call should guarantee in-order writes.
Thanks go to Patrik Lindgren and Stefan Walkner for reporting this
issue and helping to track down the source of the problem!