Fixed endless read with readexactly method.
Assume that read called as read(10, None, True).
Assume that 5 bytes is already read and TerminalSizeChanged raised.
After checking that recv_buf[0] is instance of Exception and some data is
received we should break **outer** while loop to get IncompleteReadError
raised.

Also added a unit test to validate this fix.