Add option to not close stdin/stdout/stderr on EOF from channel
This commit adds a new "recv_eof" argument to the SSHProcess redirect
feature to control whether an EOF received on the channel will cause any
output redirect targets to be closed. When this is set to false, a
caller can redirect output from multiple commands to the same target, as
well as intermix their own output to these targets in between redirects.
This is meant to parallel the current "send_eof" argument which controls
whether receiving EOF from a redirected input source triggers an EOF to
be sent to the channel or not.

This commit also fixes a mistake in one of the async file redirect
tests.