Work around a bug seen in a Huawei SFTP server
This commit attempts to work around a problem where a Huawei SFTP
server is including FILEXFER_ATTR_MODIFYTIME in the returned flags
in the response to a stat() call after negotiating SFTP version 3.
This flag is only defined in SFTPv4 and later. Thanks go to GitHub
user nielsvanhooy for reporting this amd helping to debug the issue.