Add flexibility on naming of dynamically loaded libraries
This commit adds some flexibility to the allowed filename for the Nettle
and OQS libraries. This is primarily useful on Windows, where the
installed DLL might be named something like nettle.dll, or
libnettle.dll. Also, for backward compatibility, the code also looks for
libnettle-6.dll, the only name previously supported on Windows.