Handle case where filesystem enforces UTF-8 filename requirement
Depending on configuation, ZFS volumes may sometimes enforce that
filenames be valid UTF-8 strings. This commit avoids trigger a test
failure when this is the case. Previously, this test was expecting
an error to be raised in this case, but not from the OS. Thanks
go to Robert Schütz and Martin Weinelt for reporting this issue.