Fix an issue with host-based auth when using proxy_command
This commit fixes an issue where the client host for host-based auth
can't be determined when proxy_command is used, or in other cases
where sockname information isn't available on the outbound connection.

Host based auth can still be used in this case, but only if client_host
is specified explicitly when opening the connection.

Thanks go to GitHub user yuqingm7 for reporting this issue.