Work around OpenSSH issue with RSA SHA-2 keys signed by certificates
OpenSSH versions before 7.8 didn't support RSA SHA-2
signature names in certificate key types, requiring the
use of ssh-rsa-cert-v01@openssh.com as the key type even
when using SHA-2 signatures. However, OpenSSL 8.8 and
later reject ssh-rsa-cert-v01@openssh.com as a key type
by default, requiring that the RSA SHA-2 version of the key
type be used. This makes it difficult to use RSA keys with
certificates without knowing the version of the remote
server and which key types it will accept.

This ccommit works around this by trying multiple key types
during public key and host-based authentication when using
SHA-2 signatures with RSA keys signed by certificates.

This replaces a previous change made in #517 which attempted
to fix this but only worked for OpenSSH 7.8 and later.