Add support for an "accept handler" in connection forwarding
This commit adds support for a new accept_handler argument in the
forward_local_port and forward_local_port_to_path methods in
SSHClientConnection and the ability to return an accept handler
in the server_requested method in SSHServer. This method receives
the original host & port of the incoming forwarded connection and
can return a bool to determine whether forwarding is allowed or not.
Thanks go to GitHub user zgxkbtl for suggesting this feature!