Add get_addresses() and get_port() to SSHAcceptor class
This commit adds functions to get listening socket information on
SSH listeners. New method get_addresses() gets a list of all of the
IP addresses and ports being listened on, and get_port() will return
just the listening port, but only when there's a single listening
socket or if all the sockets are listening on the same port. Otherwise,
a value of 0 is returned. Thanks go to Allison Karlitskaya for
suggesting this feature.