Treat ProxyJump as a list not a single value
The manpages and spec for OpenSSH treat ProxyJump as a list of 0 or more
proxies to jump through, however the project here currently assumes
exactly one value. This commit updates the config and tunnel logic to
connect via multiple hosts, while still allowing for str values as users
currently can manually specify a string value in a variety of cases.

fixes #635