Add an example of using "await" instead of "async with"
This commit adds an example of using "await" instead of "async with"
when opening a new SSHClientConnection. This is useful when you have
code which wants to open a connection and return it without using it
right away. When using "async with", the connection is automatically
closed when the block is exited. Thanks go to Michael Davis for
suggesting this added documentation.