Add support for multiple tunnel/ProxyJump hosts
This commit adds the ability to specify a comma-separate list of tunnel
hosts to establish a connection through, either in the "tunnel" argument
passed to the main connect and listen functions or via the ProxyJump
config file option. Thanks go to Adam Martin for suggesting this
enhancement and proposing a solution.