Add async "construct" class method to SSHConnectionOptions
This commit adds a new async construct method to crease new instances of
SSHClientConnectionOptions and SSHServerConnectionOptions. As before,
the objects are created in an executor to not block the asyncio event
loop, but this new method allows the developer to have easy access to
using an executor when constructing their own instances of these objects.