Add support for passphrase to be an awaitable
This commit expands the passphrase argument in SSHClientConnectionOptions
and SSHServerConnectionOptions to accept an awaitable, in addition to the
callable which was previously added. Options evaluation is still done in
a separate executor thread to avoid blocking the asyncio event loop, but
it's now possible to pass in an awaitable which will run as a task in the
event loop to determine the passphrase to use if you need to do "blocking"
async operations to determine this value.