Begin using Self as return type for __aenter__ methods
This commit bumps the minimum version of typing_extensions to 4.0.0 to
allow the use of "Self" as a type annotation, and changes all the places
where __aenter__ is defined to use this annotation. Thanks go to
Matthew Bradbury for suggesting this and providing a PR for this
change in SSHProcess. This commit is a superset of that PR.