Add support for OpenSSH lsetstat SFTP extension
This commit adds both client and server support for the OpenSSH "letstat"
extension to set attributes on symbolic links on platforms which support it.

This commit also adds a "follow_symlinks" optional argument on the
SFTPClient stat() and setstat() methods, as well as chown(), chmod(),
and utime(), similar to what's supported by the `os` module.

Thanks go to GitHub user eyalgolan1337 for reporting this issue!