Fix flow control for AsyncFileWriter and StreamWriter
This commit is a follow up to f2020ed, adding proper back pressure when
output is redirected to an AsyncFileWriter or StreamWriter and data is
arriving on the SSH channel faster than these writers can consume it.
Once the queue of outstanding data begins to grow, reading from the
SSH channel will be paused to allow the queue to drain somewhat before
continuing, limiting the amount of memory needed.