Further improve handling of task cancellation in SCP
This commit is a follow-on to commit 676534b to better support
KeyboardInterrupt or other forms of task cancellation during a
large SCP transfer.

This fix could also apply in other cases where `close()` or `abort()`
are called on a channel when the peer either delays or fails to send
back a MSG_CHANNEL_CLOSE, leaving the channel half-open. With this
fix, a client can close a channel and immediately wake up any blocked
tasks associated with that channel, without waiting to receive a close
message from the peer.