Make duplicate host key types an error without send_server_host_keys
Prior to the last commit, configuring multiple server host keys of the
same type raised an exception, as these keys would never be used in the
SSH handshake. This commit restores that behacvior when the new
send_server_host_keys option is disabled. Now, multiple server host keys
of the same type is only allowed when this new option is enabled. The
additional keys will still not be used in the handshake, but they will
be reported to clients as reserved keys which should be trusted for
future key rotation.

This commit also adds a note in the documentation for the new
send_server_host_keys option describing this behavior.