Add default for SFTPClient.listdir
At runtime, it's possible to call `listdir()` (i.e. without arguments)
and get a `Sequence[str]` back, due to the default argument in the
implementation. However, there is no `@overload` that contains 0
arguments, so mypy doesn't think that it's possible.

This adds the presence of the default to the sequence-of-strings
returning overload, to match the implementation