Add support for setting GSS credential store per connection
This commit adds support for setting the credential cache to use when
performing GSS authentication. While this was previously settable via
an environment variable, that didn't allow different credential caches
to be configured on a per-connection basis, since all tasks share the
same environment.

Thanks go to GitHub user zarganum for suggesting this feature and
proposing a detailed design.