Relax padding length check on OpenSSH private keys
This commit relaxes a check on the length of the padding in OpenSSH
format private keys. The padding is only supposed to be large enough to
get to the next multiple of the cipher block size, or a block size of 8
for unencrypted keys. However, PuTTYgen seems to get this wrong and use
a larger padding size on unencrypted keys, preventing AsyncSSH from
loading some unencrypted keys generated by PuTTYgen. With this change,
larger amounts of padding are allowed (up to 255 bytes), which should
help avoid this issue with PuTTYgen.