Use an executor to prevent GSSAPI calls from blocking the event loop
Some operations such as GSSAPI calls can sometimes block the event
loop if not run in an executor. However, doing that requires packet
handlers to be asynchronous. This commit adds support for async
packet handlers for key exchange and auth, and changes the GSSAPI
handlers to run the step() call in an executor.