Allow additional flexibility in specifying environment values
This commit allows the 'env' option to accept sequences of either
'key=value' strings or (key, value) tuples, in addition to accepting
a dict. The 'name=value' syntax was previously required to be sent
as a list.

As part of a previous commit, keys and values can now be either
Unicode strings or byte strings, where Unicode strings will be
encoded as UTF-8 in channel requests.