Avoid sending "MAC verification failed" error on connection close
This commit prevents AsyncSSH from sending a disconnect message with
the error "MAC verification failed" after an application-initiated
connection close. Prior to this fix, a "Disconnected by application"
message was correctly sent on connection close, but it was followed up
by a second disconnect with text "MAC verification failed" which
should not have been sent.