Improve SCP exception handling when canceling transfers in progress
This commit improves exception handling triggered when a transfer is
cancelled while in progress, either by a timeout or as a result of a
user triggering a KeyboardInterrupt.

As part of this, an "Unknown exception" should no longer be logged
for SCP server tasks which are cancelled. SCP client tasks should
also no longer raise exceptions from a timeout or KeyboardInterrupt
which is properly caught out at the application level.