Fix edge case with parallel reader
This commit fixes an issue where the parallel I/O mechanism wasn't
selected when the application set an explicit block size larger than
the requested read size, but also larger than the max read length the
server supports, so only a partial result was returned.

With this fix, the parallel I/O will always be used when the read size
is larger than what the server supports, unless the application sets
the block_size to 0 to disable it.

In the case where the application block size is larger than the max
supported read length on the server, reads will be done with the
application-requested block size, but the parallel I/O code will
compensate if the returned data is smaller than the requested size.

The application can still force a single read call with an arbitrary
size by setting block_size to 0, but it will need to deal with a
possible partial result.