Change compression to be disabled by default
This commit changes compression to default to off when AsyncSSH is
acting as a client, only enabling it if the server requires it. This
matches the OpenSSH default and avoids a potential bottleneck where
running with compression enabled can actually slow down performance
signficantly on high-bandwidth connections.

This also cleans up some unintended cosmetic differences between the
documentation for client and server connection options.