Fix issues related to supporting multiple trusted security keys
This commit better handles a use case where a client might be capable
of authenticating with one of  multiple security keys, all of which
are accepted by the server but only some of which are plugged into the
client system.

With this change, keys in client_keys which correspond to security keys
that aren't plugged into the system are ignored. Only keys which are
connected, present in client_keys, and trusted by the server will
actually be used for signing, possibly triggering a touch requirement.

Thanks go to GitHub user zanda8893 for reporting the issue and helping
to work out the details of the problem!