Add support for WebAuthN authentication with U2F security keys
This commit adds support for a new WebAuthN signature algorithm for
U2F security keys, allowing non-admin Windows users to use these
keys for authentication. Previously, U2F security keys worked on
Windows, but only when the user was an Administrator.

This new WebAuthN signature algorithm will be selected automatically
when running as a non-admin on Windows. It can also be requested
explicitly. For it to work, though, the server needs to be configured
to accept this signature algorithm. At the moment, it is disabled by
default on OpenSSH. Here's the necessary config to enable it:

PubkeyAcceptedAlgorithms=+webauthn-sk-ecdsa-sha2-nistp256@openssh.com