Fix a couple of SFTP block size issues and a couple of doc errors
This commit fixes an issue with AsyncSSH not honoring the block_size in
calls to the high level SFTP get/put/copy functions when the default
block size is smaller than the requested value. It also, fixes a problem
when a block_size of 0 is passed in to these functions. Now, a block_size
of 0 is treated the same as leaving it set to the default, automatically
choosing a value based on the advertised server limits, falling back to
16 KB if the server doesn't advertise limits.

This commit also fixes a copy of errors in the documentation for the
SFTP put() and copy() methods.

Thanks go to Krzysztof Kotlenga for finding and reporting these issues!