Add support for hostname canonicalization
This commit adds support for hostname canonicalization. It supports the
same config items used in OpenSSH config files, but also provides
options which can be passed directly, without the need to use a config
file. This commit also adds automatic reloading of the config file when
canonicalization is performed, as well as support for match keywords
"canonical" and "final".

This commit also adds support for negation of match keywords. Thanks go
to GitHub user commonism who suggested adding this support and provided
a proposed implementation.