Update hostname canonicalization to better match OpenSSH
This commit changes the initial hostname check in host canonicalization
to only look for literal IP addresses, rather than attempting to resolve
the host as if it were already fully qualified. This is more consistent
with OpenSSH and also should avoid a potentially slow additional name
lookup on every connection.