Add client and server support for SFTP copy-data extension
This commit adds client and server support for the SFTP "copy-data"
extension, and a new remote_copy() method on SFTPClient wihch allows you
to make a request to copy bytes between two files on the remote server
without needing to download and re-upload the data, if the server
supports it.

Thanks go to Ali Khosravi for suggesting this addition.