Revert part of remote_copy change to allow pathnames
This commit backs out some of the prior change to _SFTPFileCopier
so that it still handles file opens prior to deciding whether to
use remote_copy() or not. This seems to make a difference on
Windows in particular to properly handle errors in a copy.

Passing in either an open file or pathname is still supported. This
change only changes where the file open happens in the case where
_SFTPFileCopier is used.