Add support for selecting to only allow remote copy on SFTP
This commit adds a new "remote_only" argument to the SFTPClient copy()
and mcopy() functions to request that the operation only be performed
if it can be done using the "remote copy" feature. It also adds a
"supports_remote_copy" property to SFTPClient for an application to
test if the connected SFTP server supports this function.