Adjust remote_only argument to only apply to copying plain files
This commit adjusts the behavior of the "remote_only" argument in
SFTPClient copy() and mcopy() to only raise an error when attempting
to copy a plain file on a system which doesn't support remote copy.
Other operations during the copy such as creating directories and
symlinks and setting attributes will always be allowed.

When an error handler is set, it will be called once for each plain
file copy which is attempted when remote_only is set and the server
doesn't support remote copy. The exceptions will include the source
and destination pathname, to allow other forms of copy to be
performed.