Change default on SFTP max_requests to avoid excessive memory usage
This commit changes the way the default number of parallel SFTP requests
is determined. Instead of always defaulting to 128, lower values are
used for large block sizes, to reduce memory usage. With larger block
sizes, there's no need for as many parallel requests to keep the pipe
full. The minimum default is now 16, for block sizes of 256 KB or more.
The maximum default is 128, for block sizes of 32 KB or below.