Drop testing on Python 3.7 and add testing on Python 3.13
With cryptography version 44.0 including deprecation warnings for Python
3.7 with plans to drop support in its next release, this commit updates
the test matrix to drop Python 3.7 and add Python 3.13.

This commit also fixes a few issues that caused failures on Windows with
Python 3.8 and later. With these changes, Windows tests should be able
to run on all currently supported versions of Python.

This commit also updates the supported versions in the project metadata.