Add sparse file support in SFTP
This commit adds support for sparse file copying in SFTP on macOS and
Linux. Windows support may be added later.

Currently, put() is supported with any SFTP server, but get() and copy()
require an SFTP extension to request remote file data ranges, so these
currently need AsyncSSH as both the client and server. This could change
if the required extension can be standardized.