Allow the SSHServer auth_completed method to be a coroutine
This commit allows to auth_completed callback in SSHServer to be either
a callable or a coroutine, allowing async operations to be performed
from it when auth completes successfully. This callback will run to
completion before any sessions are started up, and before an acceptor
task is started if one is specified.