Fix a few issues in Windows sparse file support
This commit adds code to enable space file support on any file opened
for writing. This is needed for the put() and copy() operations arriving
on an SFTP server running in Windows to create sparse files.

This commit also fixes a minor bug with returned range results on UNIX
systems when querying for ranges in less than the full file.

Also, this commit improves the sparse file unit tests to compare ranges
created in the test rather than reading and comparing the full file
bytes (including all the holes), making the test much less expensive.