Improve handling of exceptions in connection_lost() callbacks
This commit adds wrappers around calls to connection_lost() in the
SSHClient, SSHServer, SSHClientSession, and SSHServerSession classes
to make sure that cleanup completes properly in the SSHConnection and
SSHChannel classes. Exceptions in connection_lost() will be reported
in the debug log but cleanup operations will continue, ignoring those
exceptions. Thanks go to Danil Slinchuk for reporting this issue!