Add missing awaitable return types
This commit updates the typing on the return types for the server
callback functions to reflect cases where a coroutine which returns
a session is accepted alongside a callable which returns a session
or just directly returning a session. Thanks go to Nate Hardison
for suggesting a fix for this.