Add support for forwarding incoming sessions over an SSH tunnel
This commit makes it easier for applications to accept incoming tunneled
session, connection, or TUN/TAP requests on an SSHServerConnection and
forward them over an upstream SSHClientConnection. The methods on
SSHServer to accept these requests can now return a SSHClientConnection
object to forward the traffic over, instead of having to accept the
request, open a corresponding upstream session, and then relay data
between the two SSH sessions.